Fixed --spawned-functions assertion failure
The failure was triggered when a goto-binary contained a call to
pthread_create, where the third argument to pthread_create is not a
straight function pointer.

Dealing with this is a pain, since we can't know in advance what forms
the third argument will take. I fixed the assertion failure by
returning an empty string and then checking for it. A proper fix would
be to see what the argument to pthread_create actually looks like for
real Debian packages and then to account for those.

This bug was triggered on ~15 packages out of ~1200.