Support recursive calls of functions marked always_inline
Also fix the spurious recursion resulting from asm renaming: such
functions aren't necessarily recursive as the implementation of the
asm-aliased function may be very different. In such cases just retain
the declaration of the asm-renamed symbol, but drop the
always_inline-marked function body (after having it inlined as needed).

glibc heavily uses the above with functions such as open or read, which
are replaced by variadic functions and functions with bounds checks,
respectively.