Remove composite object from the constant propagator when a field is updated
Otherwise e.g. in "memset2" we had the whole object being initialised with "{0, 0}" (which is
stored in const prop and then expanded to individual fields), then later any operation that needed
the whole-struct representation (such as the memset-of-one-field in that test, which expanded to
"whole_struct = byte_update(whole_struct, offset, int, 0)") would get the whole-struct definition
stored in const-prop, disregarding any intermediate updates that were able to directly reference
a single field.

Alternatively we could suppress ever storing composites that have an expanded representation in
the constant propagator.