Don't treat a dead instruction as a shared write
Doing so introduces an extra (nondet) generation of the symbol, which the test
anonymous-java.lang.thread treats as an opportunity to fail its test by hypothesising
that the main thread ends (and kills the temporary instance of A that the variable
under test belongs to) before the worker thread starts.

Arguably this is wrong -- the variable is stack-allocated, so either we should join all
threads before returning from "main" or we shouldn't use stack allocation for Java objects
at all -- but this simply restores symex's behaviour before field sensitivity was added,
and thus is hopefully uncontroversial.