More constant propagation
1) Within goto-symex, propagation of nondet_symbol_exprt is actually safe, as
they are instantiations of side_effect_expr_nondett and the value of these
instantiations cannot change.
2) This highlights that we also need to consider further expressions as possibly
being constant in is_constantt which we would otherwise simplify away when they
are constant (but cannot any longer do so when nondet_symbol_exprt are contained
in there).
3) Constant propagation now yields different right-hand sides in a test.
4) goto-symex applies filters on what is considered "constant" to keep the size
of expressions contained as (repeated) simplification of large expressions is
expensive. We may revisit this in future as we improve the simplifier.