Move core of constant propagation domain to util/
This part of the constant propagator has no dependency on code outside
util and enables stateful expression simplification. Moving to util will
enable future use within language front-ends. This is only a
refactoring, with the exception of the output function, which now uses
format() instead of from_expr(), which required a change to the expected
test output in one case.