Encode integer addresses together with pointer object/offset
Instead of encoding pointers with the same bit width a pointer on a
given platform has, just widen the bit-blasted encoding to include both
the previous object/offset encoding as well as an (integer) address. The
encoding is thus also trivially extended to handle larger numbers of
objects and offsets of the same width as the address.

Furthermore clean up the code to encapsulate encoding properly, and make
in-code layout of pointer encoding more natural (it's now object,
offset, integer-address).

Fixes: #436
Fixes: #311
Fixes: #94