bv_utilst::equal: stop early when the result will be false
There is no need to create any further bit-level equalities when the
result will be false anyway. On aws-c-common/aws_hash_iter_next, this
reduces the number of Boolean variables from 2291304 to 584628, and the
time spent in CaDiCaL from 191 seconds to 22 seconds.

When running on various proofs done for AWS open-source projects, this
changes the performance as follows: with CaDiCaL as back-end, the total
solver time for the hardest 46 proofs changes from XXX to XXX
seconds (XXX seconds slow-down); with Minisat, however, the hardest 49
proofs take XXX instead of XXX seconds (XX seconds
speed-up). Across these benchmarks, XXX% of variables and XXX% of
clauses are removed.